chore(virtctl): Move restart command to own file
